# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rasheed - Finance Savings Agent",
...
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "إنشاء حساب جديد",
        "description": "ينشئ حسابًا جديدًا ويرجّع زوج التوكنات (`access_token` و`refresh_token`) مع بيانات المستخدم. رقم الهاتف فريد، وكلمة المرور 8 أحرف على الأقل.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "409": {
            "description": "رقم الهاتف مسجّل مسبقًا.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "تسجيل الدخول",
        "description": "يتحقق من رقم الهاتف وكلمة المرور ويرجّع زوج التوكنات مع بيانات المستخدم.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "رقم الهاتف أو كلمة المرور غير صحيحة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent/chat": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "محادثة مع رشيد",
        "description": "يرسل رسالة المستخدم لرشيد ويرجّع رده بشخصيته المختارة، مع آخر 20 رسالة كسياق للمحادثة. قد يولّد رشيد أثناء الرد اقتراحات معلّقة (هدف جديد، تسجيل دخل...) تظهر لاحقًا في `GET /agent/actions` ولا تُنفَّذ إلا بتأكيد صريح. عند ازدحام خدمة الذكاء الاصطناعي يرجع الرد 200 مع رسالة اعتذار نصية داخل `reply`.",
        "operationId": "chat_with_agent_api_v1_agent_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "المستخدم لم يُكمل الـ Onboarding ولم يختر شخصية لرشيد بعد.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/nudge": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "رسالة مبادِرة من رشيد (Nudge)",
        "description": "يستدعيه الموبايل عند فتح التطبيق أو شاشة المحادثة. يرجّع `nudge: null` عندما لا داعي لرسالة (رد طبيعي وليس خطأ). تُولَّد الرسالة فقط عندما يكون مزاج رشيد «متحمسًا» أو «قلقًا» (حسب مصاريف المستخدم وتقدّم هدفه)، وبفاصل 12 ساعة على الأقل بين رسالتين.",
        "operationId": "get_nudge_api_v1_agent_nudge_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NudgeOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/actions": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "الاقتراحات المعلّقة",
        "description": "يرجّع اقتراحات رشيد التي لم يرد عليها المستخدم بعد (`pending`)، ويعرضها الموبايل كبطاقات موافقة/رفض. محتوى `payload` يختلف حسب `action_type`.",
        "operationId": "list_pending_actions_api_v1_agent_actions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentActionOut"
                  },
                  "type": "array",
                  "title": "Response List Pending Actions Api V1 Agent Actions Get"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/actions/check": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "توليد اقتراحات جديدة",
        "description": "يفحص وضع المستخدم ويولّد اقتراحات جديدة عند الحاجة: تصحيح تصنيف معاملات غير مصنّفة (بمساعدة Gemini)، أو دفعة لهدف قريب من الاكتمال. القائمة الفارغة رد طبيعي وليست خطأ. لا يُنفَّذ أي تعديل هنا؛ الاقتراحات فقط.",
        "operationId": "check_for_new_actions_api_v1_agent_actions_check_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentActionOut"
                  },
                  "type": "array",
                  "title": "Response Check For New Actions Api V1 Agent Actions Check Post"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/agent/actions/{action_id}/confirm": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "تأكيد اقتراح وتنفيذه",
        "description": "ينفّذ الاقتراح فعليًا على بيانات المستخدم بحسب `action_type` (تصحيح تصنيف، دفعة لهدف، إنشاء هدف، أو تسجيل دخل) ويحوّل حالته إلى `applied`. هذا هو المسار الوحيد الذي يعدّل فيه رشيد البيانات المالية، ولا يعمل إلا بطلب صريح من المستخدم.",
        "operationId": "confirm_action_api_v1_agent_actions__action_id__confirm_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentActionOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "الاقتراح تم الرد عليه سابقًا، أو نوعه غير مدعوم، أو بياناته تالفة.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "الاقتراح غير موجود، أو العنصر المرتبط به (معاملة أو هدف) غير موجود.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/actions/{action_id}/reject": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "رفض اقتراح",
        "description": "يحوّل حالة الاقتراح إلى `rejected` دون أي تعديل على بيانات المستخدم.",
        "operationId": "reject_action_api_v1_agent_actions__action_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentActionOut"
                }
              }
            }
          },
          "401": {
            "description": "التوكن ناقص أو غير صالح أو منتهي الصلاحية، أو المستخدم غير موجود/غير مفعّل.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "الاقتراح تم الرد عليه سابقًا.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "الاقتراح غير موجود.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentActionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "action_type": {
            "$ref": "#/components/schemas/AgentActionType"
          },
          "status": {
            "$ref": "#/components/schemas/AgentActionStatus"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload",
            "description": "تفاصيل الاقتراح، وشكلها يعتمد على `action_type`: `suggest_category_correction` → {transaction_id, new_category_id, new_category_name, transaction_note, transaction_amount}؛ `suggest_goal_contribution` → {goal_id, goal_title, amount}؛ `suggest_goal_creation` → {title, target_amount, breakdown, is_recurring}؛ `suggest_income_log` → {amount, note}."
          },
          "reasoning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning",
            "description": "سبب الاقتراح بصياغة رشيد (يعرضها التطبيق للمستخدم)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "action_type",
          "status",
          "payload",
          "reasoning"
        ],
        "title": "AgentActionOut",
        "examples": [
          {
            "action_type": "suggest_goal_contribution",
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "payload": {
              "amount": 18,
              "goal_id": "9c1d2e4a-7b3f-4d58-a1c6-0e5f8b2d7a91",
              "goal_title": "لابتوب جديد"
            },
            "reasoning": "إنت قريب من هدف 'لابتوب جديد' (85%)! دفعة بسيطة بتقرّبك أكتر.",
            "status": "pending"
          }
        ]
      },
      "AgentActionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "applied",
          "rejected"
        ],
        "title": "AgentActionStatus"
      },
      "AgentActionType": {
        "type": "string",
        "enum": [
          "suggest_category_correction",
          "suggest_budget_adjustment",
          "suggest_goal_contribution",
          "suggest_goal_creation",
          "suggest_income_log"
        ],
        "title": "AgentActionType"
      },
...
      "ChatMessageCreate": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Message",
            "description": "نص رسالة المستخدم لرشيد"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatMessageCreate",
        "examples": [
          {
            "message": "كم باقيلي عشان أوصل لهدفي؟"
          }
        ]
      },
      "ChatMessageOut": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply",
            "description": "رد رشيد (نص عادي، بدون أي وسوم داخلية)"
          }
        },
        "type": "object",
        "required": [
          "reply"
        ],
        "title": "ChatMessageOut",
        "examples": [
          {
            "reply": "باقيلك 90 دينار على اللابتوب، كمل بنفس الوتيرة!"
          }
        ]
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "رسالة الخطأ (بالعربي)"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "شكل الخطأ القياسي الراجع من كل الـ Endpoints.",
        "examples": [
          {
            "detail": "الهدف غير موجود"
          }
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IncomeType": {
        "type": "string",
        "enum": [
          "fixed",
          "variable"
        ],
        "title": "IncomeType",
        "description": "نوع الدخل: مهم جدًا لرشيد ليفهم السياق قبل ما ينبّه المستخدم.\nمستخدم دخله متغير ما لازم يُقارن بنفس معايير مستخدم دخله ثابت."
      },
      "NudgeOut": {
        "properties": {
          "nudge": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nudge",
            "description": "رسالة رشيد المبادِرة، أو null إذا لا داعي لأي رسالة الآن"
          }
        },
        "type": "object",
        "required": [
          "nudge"
        ],
        "title": "NudgeOut",
        "examples": [
          {
            "nudge": "ما شاء الله، صرت قريب من هدفك!"
          },
          {}
        ]
      },
...
      "PersonaOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "tagline": {
            "type": "string",
            "title": "Tagline"
          },
          "color_hex": {
            "type": "string",
            "title": "Color Hex"
          }
        },
        "type": "object",
        "required": [
          "id",
          "key",
          "display_name",
          "tagline",
          "color_hex"
        ],
        "title": "PersonaOut",
        "description": "ما منرجع system_prompt هون إطلاقًا — هاي تفاصيل داخلية بين رشيد وGemini،\nمو شي المفروض يشوفه المستخدم بشاشة اختيار الشخصية.",
        "examples": [
          {
            "color_hex": "#F57C00",
            "display_name": "رشيد الطاقة",
            "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
            "key": "energetic",
            "tagline": "شبابي ومرح وبيحمّسك توفّر"
          }
        ]
      },
...
      "TokenPair": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "يُرسل بالهيدر: Authorization: Bearer <access_token>"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "TokenPair"
      },
...
      "UserLogin": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "رقم الهاتف المسجَّل"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "كلمة المرور"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "password"
        ],
        "title": "UserLogin",
        "examples": [
          {
            "password": "Passw0rd!",
            "phone": "0791234567"
          }
        ]
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "income_type": {
            "$ref": "#/components/schemas/IncomeType"
          },
          "agent_name": {
            "type": "string",
            "title": "Agent Name"
          },
          "has_completed_onboarding": {
            "type": "boolean",
            "title": "Has Completed Onboarding"
          },
          "persona": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PersonaOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "الشخصية المختارة (null قبل إتمام الـ Onboarding)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "income_type",
          "agent_name",
          "has_completed_onboarding"
        ],
        "title": "UserOut",
        "examples": [
          {
            "agent_name": "رشيد",
            "has_completed_onboarding": true,
            "id": "9c1d2e4a-7b3f-4d58-a1c6-0e5f8b2d7a91",
            "income_type": "variable",
            "name": "سامر",
            "persona": {
              "color_hex": "#F57C00",
              "display_name": "رشيد الطاقة",
              "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "key": "energetic",
              "tagline": "شبابي ومرح وبيحمّسك توفّر"
            },
            "phone": "0791234567"
          }
        ]
      },
      "UserRegister": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name",
            "description": "الاسم الظاهر للمستخدم"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "minLength": 8,
            "title": "Phone",
            "description": "رقم الهاتف — فريد لكل حساب"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "8 أحرف على الأقل"
          },
          "income_type": {
            "$ref": "#/components/schemas/IncomeType",
            "description": "`fixed` دخل ثابت (راتب)، أو `variable` دخل غير منتظم (فريلانس/أجرة يومية)",
            "default": "variable"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "password"
        ],
        "title": "UserRegister",
        "examples": [
          {
            "income_type": "variable",
            "name": "سامر",
            "password": "Passw0rd!",
            "phone": "0791234567"
          }
        ]
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}